Semestre2/T3/Reconhecimento/requirements.txt
boto3==1.24.88
botocore==1.27.96
jmespath==1.0.1
numpy==2.1.1
opencv-python==4.10.0.84
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
RPi.GPIO==0.7.1
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.19